unit test and challenge test

+++ src/main/java/com/textile/challenge/pricecatalog/repository/PriceListRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PriceListRepository extends JpaRepository<PriceList, Long> {
+public interface PriceListRepository extends JpaRepository<PriceList, Integer> {
 }
