+++ src/main/java/com/textile/challenge/pricecatalog/service/BrandService.java
@@ -4,6 +4,6 @@ import com.textile.challenge.pricecatalog.util.exceptions.BrandNotFoundException
 
 public interface BrandService {
 
-    void validateBrandId(Long brandId) throws BrandNotFoundException;
+    void validateBrandId(Integer brandId) throws BrandNotFoundException;
 
 }
